-----BEGIN PGP SIGNED MESSAGE----- Hash: SHA1
The addition of the dot_writer class (bringing in Spirit as a dependency
for the profiler) forces my hand in moving to the next stage. We've reached
the point at which the profiler has exceeded the constraints of an LLVM plugin;
Spirit cannot compile without RTTI, and the plugin cannot run with it.

-----BEGIN PGP SIGNATURE-----
Version: GnuPG v1.4.9 (GNU/Linux)

iEYEARECAAYFAkym4l4ACgkQO/fqqIuE2t5icwCgvepu9KXyUm+otnGYlMyO+doM
cowAn1+zRYdSevkrWE3NjnIzbDFxA1+T
=f1S1
-----END PGP SIGNATURE-----